refactor: organize pages folder
