+++ src/pages/Teams/__tests__/testTeams.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import {render, screen, waitForElementToBeRemoved} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import * as API from '../../api';
-import Teams from '../Teams';
+import * as API from '../../../api';
+import Teams from '../index';
 
 jest.mock('react-router-dom', () => ({
     useLocation: () => ({
