--- src/pages/UserOverview.tsx
@@ -1,38 +1,0 @@
-import * as React from 'react';
-import {useLocation} from 'react-router-dom';
-import {IUserData} from 'types';
-import Card from '../components/Card';
-import {Container} from '../components/GlobalComponents';
-import Header from '../components/Header';
-
-var mapU = (user: IUserData) => {
-    var columns = [
-        {
-            key: 'Name',
-            value: `${user.firstName} ${user.lastName}`,
-        },
-        {
-            key: 'Display Name',
-            value: user.displayName,
-        },
-        {
-            key: 'Location',
-            value: user.location,
-        },
-    ];
-    return <Card columns={columns} hasNavigation={false} navigationProps={user} />;
-};
-
-const UserOverview = () => {
-    const location = useLocation();
-    return (
-        <Container>
-            <Header
-                title={`User ${location.state.firstName} ${location.state.lastName}`}
-            />
-            {mapU(location.state)}
-        </Container>
-    );
-};
-
-export default UserOverview;
